av1_cx_iface.c: fix signed/unsigned warning
In Visual Studio / cl.exe builds after:
 32c4161373 Use g_bit_depth during input validation

```
av1\av1_cx_iface.c(805,30):
  warning C4018: '>': signed/unsigned mismatch
```

+++ av1/av1_cx_iface.c
@@ -802,7 +802,7 @@ static aom_codec_err_t validate_config(aom_codec_alg_priv_t *ctx,
   RANGE_CHECK_HI(extra_cfg, cq_level, 63);
   RANGE_CHECK(cfg, g_bit_depth, AOM_BITS_8, AOM_BITS_12);
   RANGE_CHECK(cfg, g_input_bit_depth, 8, 12);
-  if (cfg->g_input_bit_depth > cfg->g_bit_depth)
+  if (cfg->g_input_bit_depth > (unsigned int)cfg->g_bit_depth)
     ERROR("Input bit-depth must not exceed codec bit-depth");
   RANGE_CHECK(extra_cfg, content, AOM_CONTENT_DEFAULT, AOM_CONTENT_INVALID - 1);
